fix(64bit): resolve the 16 -Wshorten-64-to-32 LP64 truncations (#14)
Phase 6, commit 2. Make every -Wshorten-64-to-32 site explicit so golden
stays byte-identical (the implicit conversion already truncated exactly
this way; the cast/wider-type only documents it).

- olympia/z.c + mapgen/z.c readlin path: `nread` retyped int->ssize_t
  (its source is read()), clearing both `nread = read(...)` assignment
  sites in each; downstream indexing/compares are unaffected.
- mapgen/z.c str_save: (unsigned) cast on `strlen(s) + 1` feeding the
  mapgen my_malloc(unsigned size). (olympia's my_malloc is checked_alloc
  with a size_t arg, so its str_save was never flagged.)
- olympia/code.c letter_val: `return (int)(p-let)` — index into a fixed
  short string.
- mapgen/island.c: `x_size = (int) strlen(...)` — map row length.
- olympia/add.c: `rnd(1, (int) strlen(symbols))` — password symbol set.
- strlen()->int name/line/word lengths, provably <2^31: documented (int)
  casts in z.c/mapgen z.c fuzzy_strcmp, c2.c line_length_check, check.c
  check_loc_name_lengths, eat.c do_eat_command, report.c
  strip_leading_stupid_word.

G3-specific vs G2: add.c (sbaillie password feature) and island.c are
extra sites with no G2 precedent. md5_int's `return buf[0]` needed NO
change here — G3's word32 is uint32_t (the old `unsigned long` typedef is
commented out), so buf[0]->int is a 32->32 conversion, unlike G2 where
word32 was unsigned long and the cast was required. 16 sites total.

Debug and asan-ubsan golden gates both YES (byte-identical); asan/ubsan
clean.

+++ mapgen/island.c
@@ -179,7 +179,7 @@ int main(int argc, char *argv[])
 		if (sscanf(buffer, "%[^\r\n]", map[y_size]))
 		{
 			if (!x_size)
-				x_size = strlen(map[y_size]);
+				x_size = (int) strlen(map[y_size]);	/* map row len fits 32 bits */
 			if (strlen(map[y_size]) == x_size)
 				y_size++;
 		}
